refactor: move bot handler tests to dedicated modules, create test_helpers module
